+++ .eslintrc.js
@@ -61,7 +61,7 @@ module.exports = {
         'plugin:react/jsx-runtime',
         'prettier',
       ],
-      plugins: ['promise', 'sonarjs', 'unicorn', 'sonar', '@shopify'],
+      plugins: ['promise', 'sonarjs', 'unicorn', '@shopify'],
       globals: {
         atom: false,
         document: false,
@@ -337,8 +337,9 @@ module.exports = {
         '@typescript-eslint/no-unused-expressions': 'error',
         'sonarjs/no-small-switch': 'error',
         'sonarjs/no-duplicated-branches': 'error',
-        'sonar/prefer-promise-shorthand': 'error',
-        'sonar/no-dead-store': 'error',
+        'sonarjs/prefer-promise-shorthand': 'error',
+        'sonarjs/no-dead-store': 'error',
+        'sonarjs/void-use': 'error',
         'unicorn/prefer-node-protocol': 'error',
         'import-x/no-unresolved': [
           'error',
@@ -375,6 +376,7 @@ module.exports = {
       excludedFiles: ['**/*.{md,mdx}/*.{ts,tsx}'],
       // extends: ['plugin:@typescript-eslint/recommended-type-checked'],
       rules: {
+        // '@typescript-eslint/no-redundant-type-constituents': 'error',
         '@typescript-eslint/prefer-optional-chain': 'error',
         '@typescript-eslint/no-unnecessary-type-assertion': 'error',
         '@typescript-eslint/no-floating-promises': 'error',
@@ -397,11 +399,10 @@ module.exports = {
         projectService: {
           allowDefaultProject: [
             'examples/monaco-graphql-react-vite/vite.config.ts',
-            'packages/*/vitest.config.mts',
-            'packages/*/vite.config.mts',
+            'packages/{graphiql,graphiql-plugin-explorer,graphiql-plugin-code-exporter}/vite.config.mts',
+            'packages/{codemirror-graphql,graphiql-toolkit,graphql-language-service-cli,graphql-language-service,monaco-graphql,vscode-graphql-syntax,graphiql}/vitest.config.mts',
 
             'packages/cm6-graphql/__tests__/test.spec.ts',
-            'packages/graphiql-react/setup-files.ts',
             'packages/graphiql/src/GraphiQL.spec.tsx',
             'packages/vscode-graphql-syntax/tests/*.spec.ts',
             'packages/graphql-language-service-cli/src/__tests__/*.test.ts',
@@ -492,7 +493,7 @@ module.exports = {
       },
     },
     {
-      // Rule prefer await to then without React packages because it's ugly to have `async IIFE` inside `useEffect`
+      // Rule to prefer await to then without React packages because it's ugly to have `async IIFE` inside `useEffect`
       files: ['packages/**'],
       excludedFiles: ['packages/graphiql/**', 'packages/graphiql-react/**'],
       rules: {
@@ -500,7 +501,7 @@ module.exports = {
       },
     },
     {
-      files: ['packages/{graphiql-react,graphiql}/**'],
+      files: ['packages/{graphiql-react,graphiql}/**/*.{ts,tsx}'],
       rules: {
         '@typescript-eslint/no-restricted-imports': [
           'error',
@@ -511,6 +512,7 @@ module.exports = {
           },
         ],
         'react-hooks/react-compiler': 'error',
+        '@typescript-eslint/no-deprecated': 'error',
       },
     },
     {
@@ -539,6 +541,12 @@ module.exports = {
       },
     },
     {
+      files: ['**/*.d.ts'],
+      rules: {
+        'no-var': 'off',
+      },
+    },
+    {
       // ❗ALWAYS LAST
       // Rules for codeblocks inside Markdown/MDX
       files: ['**/*.{md,mdx}/*.{js,jsx,ts,tsx}'],
@@ -550,7 +558,7 @@ module.exports = {
         'no-undef': 'off',
         'react/jsx-no-undef': 'off',
         'react-hooks/rules-of-hooks': 'off',
-        'sonar/no-dead-store': 'off',
+        'sonarjs/no-dead-store': 'off',
         '@typescript-eslint/no-restricted-imports': 'off',
       },
     },