+++ packages/graphiql-react/src/editor/query-editor.ts
@@ -11,7 +11,7 @@ import {
   GraphQLDocumentMode,
   OperationFacts,
 } from 'graphql-language-service';
-import { MutableRefObject, useEffect, useRef } from 'react';
+import { RefObject, useEffect, useRef } from 'react';
 
 import { useExecutionContext } from '../execution';
 import { useExplorerContext } from '../explorer';
@@ -152,7 +152,7 @@ export function useQueryEditor(
   const merge = useMergeQuery({ caller: caller || _useQueryEditor });
   const prettify = usePrettifyEditors({ caller: caller || _useQueryEditor, onPrettifyQuery });
   const ref = useRef<HTMLDivElement>(null);
-  const codeMirrorRef = useRef<CodeMirrorType>();
+  const codeMirrorRef = useRef<CodeMirrorType>(undefined);
 
   const onClickReferenceRef = useRef<
     NonNullable<UseQueryEditorArgs['onClickReference']>
@@ -477,7 +477,7 @@ export function useQueryEditor(
 function useSynchronizeSchema(
   editor: CodeMirrorEditor | null,
   schema: GraphQLSchema | null,
-  codeMirrorRef: MutableRefObject<CodeMirrorType | undefined>,
+  codeMirrorRef: RefObject<CodeMirrorType | undefined>,
 ) {
   useEffect(() => {
     if (!editor) {
@@ -496,7 +496,7 @@ function useSynchronizeSchema(
 function useSynchronizeValidationRules(
   editor: CodeMirrorEditor | null,
   validationRules: ValidationRule[] | null,
-  codeMirrorRef: MutableRefObject<CodeMirrorType | undefined>,
+  codeMirrorRef: RefObject<CodeMirrorType | undefined>,
 ) {
   useEffect(() => {
     if (!editor) {
@@ -515,7 +515,7 @@ function useSynchronizeValidationRules(
 function useSynchronizeExternalFragments(
   editor: CodeMirrorEditor | null,
   externalFragments: Map<string, FragmentDefinitionNode>,
-  codeMirrorRef: MutableRefObject<CodeMirrorType | undefined>,
+  codeMirrorRef: RefObject<CodeMirrorType | undefined>,
 ) {
   const externalFragmentList = [...externalFragments.values()]; // eslint-disable-line react-hooks/exhaustive-deps -- false positive, variable is optimized by react-compiler, no need to wrap with useMemo